refactor(browse): replace lonely-alert with shared ShareDialog offcanvas
Promote the no-results invite CTA to an app-level ShareDialog that lives
in features/app/components, driven by a delayed trigger prop so the
prompt only surfaces after the empty state has been stable for 5s. The
dismiss state is tracked in appStore.shareCtaDismissed so the user isn't
prompted again in the same session. Drops the old BAlert + NoResultsCTA
+ BModal trio (three components collapsed into one BOffcanvas).
