+++ apps/frontend/components.d.ts
@@ -30,6 +30,7 @@ declare module 'vue' {
     BFormFile: typeof import('bootstrap-vue-next/components/BFormFile')['BFormFile']
     BFormFloatingLabel: typeof import('bootstrap-vue-next/components/BForm')['BFormFloatingLabel']
     BFormGroup: typeof import('bootstrap-vue-next/components/BFormGroup')['BFormGroup']
+    BFormInput: typeof import('bootstrap-vue-next/components/BFormInput')['BFormInput']
     BFormInvalidFeedback: typeof import('bootstrap-vue-next/components/BForm')['BFormInvalidFeedback']
     BFormRadio: typeof import('bootstrap-vue-next/components/BFormRadio')['BFormRadio']
     BFormRadioGroup: typeof import('bootstrap-vue-next/components/BFormRadio')['BFormRadioGroup']
@@ -37,6 +38,7 @@ declare module 'vue' {
     BFormSelectOption: typeof import('bootstrap-vue-next/components/BFormSelect')['BFormSelectOption']
     BFormTextarea: typeof import('bootstrap-vue-next/components/BFormTextarea')['BFormTextarea']
     BInput: typeof import('bootstrap-vue-next/components/BFormInput')['BInput']
+    BInputGroup: typeof import('bootstrap-vue-next/components/BInputGroup')['BInputGroup']
     BListGroup: typeof import('bootstrap-vue-next/components/BListGroup')['BListGroup']
     BListGroupItem: typeof import('bootstrap-vue-next/components/BListGroup')['BListGroupItem']
     BModal: typeof import('bootstrap-vue-next/components/BModal')['BModal']
@@ -80,6 +82,7 @@ declare global {
   const BFormFile: typeof import('bootstrap-vue-next/components/BFormFile')['BFormFile']
   const BFormFloatingLabel: typeof import('bootstrap-vue-next/components/BForm')['BFormFloatingLabel']
   const BFormGroup: typeof import('bootstrap-vue-next/components/BFormGroup')['BFormGroup']
+  const BFormInput: typeof import('bootstrap-vue-next/components/BFormInput')['BFormInput']
   const BFormInvalidFeedback: typeof import('bootstrap-vue-next/components/BForm')['BFormInvalidFeedback']
   const BFormRadio: typeof import('bootstrap-vue-next/components/BFormRadio')['BFormRadio']
   const BFormRadioGroup: typeof import('bootstrap-vue-next/components/BFormRadio')['BFormRadioGroup']
@@ -87,6 +90,7 @@ declare global {
   const BFormSelectOption: typeof import('bootstrap-vue-next/components/BFormSelect')['BFormSelectOption']
   const BFormTextarea: typeof import('bootstrap-vue-next/components/BFormTextarea')['BFormTextarea']
   const BInput: typeof import('bootstrap-vue-next/components/BFormInput')['BInput']
+  const BInputGroup: typeof import('bootstrap-vue-next/components/BInputGroup')['BInputGroup']
   const BListGroup: typeof import('bootstrap-vue-next/components/BListGroup')['BListGroup']
   const BListGroupItem: typeof import('bootstrap-vue-next/components/BListGroup')['BListGroupItem']
   const BModal: typeof import('bootstrap-vue-next/components/BModal')['BModal']
